Added confirm for reset button

+++ src/js/main.js
@@ -272,25 +272,27 @@ clickerElement.addEventListener('mouseup', () => {
 });
 
 resetElement.addEventListener('click', () => {
-  updateScore(-score);
-  updateBank(-bank);
-  updateClickPerSeconds(-clickPerSeconds);
+  if (confirm('Click OK to reset your game. All your progress will be deleted.')) {
+    updateScore(-score);
+    updateBank(-bank);
+    updateClickPerSeconds(-clickPerSeconds);
 
-  for (const bonus of store) {
-    updateBonusCount(bonus, -bonus.count);
-    updateBonusCps(bonus, -bonus.cps);
-    updateBonusPrice(bonus, -bonus.price);
+    for (const bonus of store) {
+      updateBonusCount(bonus, -bonus.count);
+      updateBonusCps(bonus, -bonus.cps);
+      updateBonusPrice(bonus, -bonus.price);
+    }
+
+    hasBoost = false;
+    id = uuid();
+    localStorage.setItem('id', id);
+    notificationsList.innerHTML = '';
+
+    name = '';
+    setRandomNinjaName();
+
+    updateBonusAvailability();
   }
-
-  hasBoost = false;
-  id = uuid();
-  localStorage.setItem('id', id);
-  notificationsList.innerHTML = '';
-
-  name = '';
-  setRandomNinjaName();
-
-  updateBonusAvailability();
 });
 
 // DATABASE
